Dedupe `e_secret_stairs`
