+++ config/splat.us.sttop.yaml
@@ -45,7 +45,9 @@ segments:
       - [0xA68, .data, e_red_door_tiles]
       - [0xA98, .data, e_lock_camera]
       - [0xAF8, .data, e_breakable]
-      - [0xC08, data, 296F4]
+      - [0xC08, data]
+      - [0xC0C, .data, e_secret_stairs]
+      - [0xCBC, data]
       - [0xCC8, .data, e_elevator]
       - [0xCF0, data]
       - [0xD10, .data, d_prize_drops]
@@ -68,7 +70,7 @@ segments:
       - [0x288E8, .rodata, e_lock_camera]
       - [0x28904, .rodata, e_richter_fight_trigger]
       - [0x2891C, .rodata, e_breakable]
-      - [0x2893C, .rodata, 296F4]
+      - [0x2893C, .rodata, e_secret_stairs]
       - [0x28978, .rodata, 2B308]
       - [0x2898C, .rodata, cutscene]
       - [0x28A30, .rodata, post_richter_cutscene]
@@ -82,7 +84,8 @@ segments:
       - [0x28D00, c, e_lock_camera]
       - [0x29250, c, e_richter_fight_trigger]
       - [0x29390, c, e_breakable]
-      - [0x299EC, c, 296F4]
+      - [0x299EC, c, e_secret_stairs]
+      - [0x2A354, c, e_breakable_wall]
       - [0x2A624, c, e_elevator]
       - [0x2B308, c, 2B308]
       - [0x2B768, c, cutscene]
