+++ config/symbols.pspeu.sttop.txt
@@ -1,4 +1,3 @@
-CreateEntityFromCurrentEntity = 0x09238598;
 CreateEntityFromEntity = 0x09238620;
 EntityAxeKnightBlue = 0x09239580;
 EntityAxeKnightThrowingAxe = 0x0923A6F0;
@@ -23,8 +22,6 @@ EntityFleaRider = 0x09245958;
 EntityRoomForeground = 0x09246048;
 DestroyEntity = 0x09246160;
 AnimateEntity = 0x09246280;
-MoveEntity = 0x09246540;
-UnkCollisionFunc3 = 0x092465C8;
 AllocEntity = 0x09246B60;
 SetStep = 0x09247070;
 InitializeEntity = 0x09247190;
@@ -39,7 +36,8 @@ func_us_801AD320 = 0x09249208;
 func_us_801AD400 = 0x09249388;
 EntityTriangleElevator = 0x09249B38;
 func_us_801AABA4 = 0x0924A3B0;
-EntitySecretButton = 0x0924AE50;
+EntityStairSwitch = 0x0924AE50;
+func_us_801A9CB8 = 0x0924B1E8;
 EntitySecretStairs = 0x0924B4F0;
 EntityRedEyeBust = 0x0924BB48;
 GetLang = 0x0924BFB0;
@@ -81,6 +79,8 @@ D_us_80182E44 = 0x092580F0;
 TOP_Overlay = 0x09275D60;
 TOP_PrizeDrops = 0x09275E00;
 D_psp_E_BREAKABLE_DEBRIS = 0x09276BD8;
+D_psp_E_SECRET_STAIRS = 0x09276C50;
+D_psp_E_STAIR_SWITCH = 0x09276C58;
 D_us_801BC514 = 0x09276C90;
 g_Dialogue = 0x09277258; // size:0x40
 g_SkipCutscene = 0x09277298;
