+++ src/st/top/e_secret_stairs.c
@@ -1,9 +1,38 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "top.h"
-#include "sfx.h"
 
 extern EInit g_EInitTOPCommon;
-extern u16 D_us_80180CAC[];
+
+static Point16 D_us_80180C0C[] = {
+    {0x0580, 0x0330},
+    {0x0598, 0x0348},
+    {0x05B0, 0x0360},
+    {0x05C0, 0x0370},
+};
+
+// g_Tilemap.fg[i0] = i1
+static u16 D_us_80180C1C[][2] = {
+    {0x1956, 0x0645}, {0x1957, 0x0646}, {0x19D6, 0x0644}, {0x19D7, 0x0643},
+    {0x19D8, 0x0641}, {0x1A57, 0x0642}, {0x1A58, 0x0640}, {0x1A59, 0x063B},
+    {0x1AD8, 0x063F}, {0x1AD9, 0x063C}, {0x1ADA, 0x0639}, {0x1B59, 0x063E},
+    {0x1B5A, 0x063A}, {0x1B5B, 0x0638}, {0x1BDA, 0x063D}, {0x1BDB, 0x0637},
+    {0x1BDC, 0x0636}, {0xFFFF, 0xFFFF},
+};
+static u16 D_us_80180C64[][2] = {
+    {0x1956, 0x0119}, {0x1957, 0x0118}, {0x19D6, 0x0002}, {0x19D7, 0x0002},
+    {0x19D8, 0x017F}, {0x1A57, 0x0166}, {0x1A58, 0x0187}, {0x1A59, 0x0188},
+    {0x1AD8, 0x018F}, {0x1AD9, 0x0190}, {0x1ADA, 0x0191}, {0x1B59, 0x0198},
+    {0x1B5A, 0x0199}, {0x1B5B, 0x019A}, {0x1BDA, 0x01A1}, {0x1BDB, 0x01A2},
+    {0x1BDC, 0x01A3}, {0xFFFF, 0xFFFF},
+};
+static s16 D_us_80180CAC[] = {
+    0, 8, 0, 4, 4, -4, -8, 0,
+};
+
+#ifdef VERSION_PSP
+extern s32 E_ID(STAIR_SWITCH);
+extern s32 E_ID(SECRET_STAIRS);
+#endif
 
 void EntityStairSwitch(Entity* self) {
     Entity* entity;
@@ -15,13 +44,13 @@ void EntityStairSwitch(Entity* self) {
         self->hitboxWidth = 6;
         self->hitboxHeight = 6;
         self->hitboxState = 2;
-        if (self->params != 0) {
+        if (self->params) {
             self->step = 4;
             self->zPriority += 2;
         } else if (g_CastleFlags[TOP_SECRET_STAIRS]) {
             self->animCurFrame = 4;
-            self->hitboxState = 0;
             self->palette++;
+            self->hitboxState = 0;
             g_Tilemap.fg[0x1AE4] = 0x328;
             self->step = 3;
         } else {
@@ -30,13 +59,13 @@ void EntityStairSwitch(Entity* self) {
         break;
 
     case 1:
-        if (self->hitFlags != 0) {
+        if (self->hitFlags) {
             g_Tilemap.fg[0x1AE4] = 0x328;
             self->animCurFrame = 4;
             g_api.PlaySfx(SFX_WALL_DEBRIS_B);
             entity = AllocEntity(&g_Entities[224], &g_Entities[256]);
             if (entity != NULL) {
-                CreateEntityFromEntity(E_STAIR_SWITCH, self, entity);
+                CreateEntityFromEntity(E_ID(STAIR_SWITCH), self, entity);
                 entity->params = 1;
             }
             self->step++;
@@ -156,10 +185,6 @@ void func_us_801A9CB8(Entity* self) {
     }
 }
 
-extern Point16 D_us_80180C0C[];
-extern u16 D_us_80180C1C;
-extern u16 D_us_80180C64;
-
 void EntitySecretStairs(Entity* self) {
     Entity* entity;
     s32 i;
@@ -178,7 +203,7 @@ void EntitySecretStairs(Entity* self) {
             self->ext.secretStairs.unk84 = 1;
             entity = self + 1;
             for (i = 0; i < 3; i++, entity++) {
-                CreateEntityFromCurrentEntity(E_SECRET_STAIRS, entity);
+                CreateEntityFromCurrentEntity(E_ID(SECRET_STAIRS), entity);
                 entity->params = i + 1;
             }
         } else {
@@ -188,9 +213,9 @@ void EntitySecretStairs(Entity* self) {
                 self->zPriority++;
 
                 if (!g_CastleFlags[TOP_SECRET_STAIRS]) {
-                    var_a1 = &D_us_80180C64;
+                    var_a1 = *D_us_80180C64;
                 } else {
-                    var_a1 = &D_us_80180C1C;
+                    var_a1 = *D_us_80180C1C;
                 }
 
                 while (*var_a1 != 0xFFFF) {
@@ -273,12 +298,12 @@ void EntitySecretStairs(Entity* self) {
                     entity = self + 1;
                     entity->ext.secretStairs.unk84 = 1;
                 } else {
-                    var_a1 = &D_us_80180C1C;
+                    var_a1 = *D_us_80180C1C;
                     while (*var_a1 != 0xFFFF) {
                         g_Tilemap.fg[var_a1[0]] = var_a1[1];
                         var_a1 += 2;
                     }
-                    var_a1 = &D_us_80180C1C;
+                    var_a1 = *D_us_80180C1C;
                     while (*var_a1 != 0xFFFF) {
                         g_Tilemap.fg[var_a1[0]] = var_a1[1];
                         var_a1 += 2;
@@ -287,118 +312,6 @@ void EntitySecretStairs(Entity* self) {
                 self->step++;
             }
             break;
-        }
-        break;
-    }
-}
-
-// breakable wall behind leap stone or power of mist
-extern s16 D_us_80180CBC[];
-extern s16 D_us_80180CC2[];
-
-void EntityBreakableWall(Entity* self) {
-    s32 wallStatus;
-    Entity* entity;
-    s32 i;
-    s16* wallTiles;
-    s32 var_a1;
-
-    switch (self->step) {
-    case 0:
-        InitializeEntity(g_EInitTOPCommon);
-        self->zPriority = 0x6A;
-        self->animCurFrame = 0xF;
-        self->hitboxWidth = 8;
-        self->hitboxHeight = 0x18;
-        self->hitboxOffX = 8;
-        self->hitboxOffY = 0;
-        self->hitboxState = 2;
-        if (!self->params) {
-            if (g_CastleFlags[OVL_EXPORT(SECRET_WALL_1_BROKEN)]) {
-                wallStatus = 3;
-            } else {
-                wallStatus = 0;
-            }
-        } else if (g_CastleFlags[OVL_EXPORT(SECRET_WALL_2_BROKEN])) {
-            wallStatus = 3;
-        } else {
-            wallStatus = 0;
-        }
-
-        wallTiles = &D_us_80180CBC[wallStatus];
-        if (!self->params) {
-            var_a1 = 0x2518;
-        } else {
-            var_a1 = 0x3518;
-        }
-        for (i = 0; i < 3; i++, wallTiles++) {
-            g_Tilemap.fg[var_a1] = *wallTiles;
-            var_a1 += 0x80;
-        }
-        if (wallStatus != 0) {
-            DestroyEntity(self);
-            break;
-        }
-        // fallthrough
-
-    case 1:
-        if (self->hitFlags) {
-            g_api.PlaySfx(SFX_WALL_DEBRIS_B);
-            self->step++;
-        }
-        break;
-
-    case 2:
-        self->ext.breakable.breakCount++;
-        self->animCurFrame++;
-        entity = AllocEntity(&g_Entities[224], &g_Entities[256]);
-        if (entity != NULL) {
-            CreateEntityFromEntity(E_EXPLOSION, self, entity);
-            entity->params = 0x13;
-            entity->zPriority = self->zPriority + 1;
-        }
-        self->ext.breakable.resetTimer = 0x20;
-        self->step++;
-        if (self->ext.breakable.breakCount == 3) {
-            if (!self->params) {
-                g_CastleFlags[OVL_EXPORT(SECRET_WALL_1_BROKEN)] = 1;
-            } else {
-                g_CastleFlags[OVL_EXPORT(SECRET_WALL_2_BROKEN)] = 1;
-            }
-            wallTiles = D_us_80180CC2;
-            if (!self->params) {
-                var_a1 = 0x2518;
-            } else {
-                var_a1 = 0x3518;
-            }
-
-            for (i = 0; i < 3; i++, wallTiles++) {
-                g_Tilemap.fg[var_a1] = *wallTiles;
-                var_a1 += 0x80;
-            }
-
-            self->entityId = 0xC;
-            self->pfnUpdate = EntityHeartDrop;
-            self->step = 0;
-            self->step_s = 0;
-            self->pose = 0;
-            self->poseTimer = 0;
-
-            if (!self->params) {
-                // PrizeDrops[2] = Fire Mail
-                self->params = 2;
-            } else {
-                // PrizeDrops[1] = Turkey
-                self->params = 1;
-            }
-            // n.b.! odd return, but necessary for PSP
-            return;
-        }
-        break;
-
-    case 3:
-        if (--self->ext.breakable.resetTimer == 0) {
-            self->step = 1;
         }
         break;
     }
