+++ src/st/top_psp/e_lion_lamp.c
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "../top/top.h"
+
+INCLUDE_ASM("st/top_psp/nonmatchings/top_psp/e_lion_lamp", EntityRedEyeBust);
+
+INCLUDE_ASM("st/top_psp/nonmatchings/top_psp/e_lion_lamp", func_pspeu_0924BCD0);
