--- src/st/top_psp/unk_12048.c
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../top/top.h"
-
-INCLUDE_ASM("st/top_psp/nonmatchings/top_psp/unk_12048", func_pspeu_09249B38);
-
-INCLUDE_ASM("st/top_psp/nonmatchings/top_psp/unk_12048", func_pspeu_0924A3B0);
