--- src/st/top_psp/unk_137D0.c
@@ -1,12 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../top/top.h"
-
-INCLUDE_ASM("st/top_psp/nonmatchings/top_psp/unk_137D0", EntitySecretButton);
-
-INCLUDE_ASM("st/top_psp/nonmatchings/top_psp/unk_137D0", func_pspeu_0924B1E8);
-
-INCLUDE_ASM("st/top_psp/nonmatchings/top_psp/unk_137D0", EntitySecretStairs);
-
-INCLUDE_ASM("st/top_psp/nonmatchings/top_psp/unk_137D0", EntityRedEyeBust);
-
-INCLUDE_ASM("st/top_psp/nonmatchings/top_psp/unk_137D0", func_pspeu_0924BCD0);
